requirements.txt
flask 
google-cloud-vision 
Pillow